add retries to the entrypoint value

+++ src/eth_provider/constant.rs
@@ -20,7 +20,7 @@ pub const ADDRESS_HEX_STRING_LEN: usize = 40;
 /// Starknet Modulus: 0x800000000000011000000000000000000000000000000000000000000000001
 pub const STARKNET_MODULUS: U256 = U256::from_limbs([0x1, 0, 0, 0x0800_0000_0000_0011]);
 /// Maximum number of times a transaction can be retried
-pub const TRANSACTION_MAX_RETRIES: u64 = 10;
+pub const TRANSACTION_MAX_RETRIES: u8 = 10;
 
 #[cfg(feature = "hive")]
 use {
